refactor(compaction): one module per capability + shared utils
Split the monolithic _capability.py into a module per capability plus a _shared module
for cross-capability utilities (token estimation, the CompactionStrategy protocol,
tool-pair-safe cutoffs, first-user preservation, in-place tool-result clearing).

- _shared.py exposes its package-internal API without leading underscores (the module
  itself is private); genuinely module-local helpers keep their underscore.
- _sliding_window / _clear_tool_results / _deduplicate_file_reads / _limit_warner /
  _summarizing_compaction / _tiered_compaction each own one capability.
- __init__.py re-exports the public API; tests import privates from their new homes.

No behavior change. Every compaction module at 100% branch coverage; ruff + pyright strict clean.
